@@ -4,7 +4,7 @@
   "description": "snyk library and cli utility",
   "main": "dist/lib/index.js",
   "files": [
-    "help",
+    "bin",
     "dist",
     "config.default.json",
     "SECURITY.md",
@@ -18,7 +18,7 @@
     "doc": "help/commands-help"
   },
   "bin": {
-    "snyk": "dist/cli/index.js"
+    "snyk": "bin/snyk"
   },
   "engines": {
     "node": ">=10"
@@ -28,11 +28,11 @@
     "packages/*"
   ],
   "scripts": {
-    "build": "tsc --build",
+    "build": "npm run build:dev",
     "build-watch": "tsc -w",
     "find-circular": "npm run build && madge --circular ./dist",
     "check-dev-environment": "ts-node ./scripts/check-dev-environment.ts",
-    "prepare": "npm run check-dev-environment && npm run build",
+    "prepare": "npm run check-dev-environment && npx lerna run build --ignore snyk && npm run build",
     "generate-help": "./help/generator/generate-docs.sh",
     "test:common": "npm run check-tests && npm run lint",
     "test:acceptance": "tap test/acceptance/**/*.test.* test/acceptance/*.test.* -Rspec --timeout=300 --node-arg=-r --node-arg=ts-node/register",
@@ -54,8 +54,11 @@
     "format": "prettier --write '**/*.{js,ts,json,yaml,yml,md}'",
     "format:changes": "./scripts/format/prettier-changes.sh",
     "check-tests": "! grep 'test\\.only' test/*.test.js -n",
-    "snyk-auth": "node --require ts-node/register src/cli auth $SNYK_API_KEY",
-    "snyk-auth-windows": "bash ./scripts/auth-snyk-win.sh"
+    "snyk-auth": "node dist/cli auth $SNYK_API_KEY",
+    "snyk-auth-windows": "bash ./scripts/auth-snyk-win.sh",
+    "build:dev": "webpack --mode=development",
+    "build:prod": "webpack --mode=production",
+    "watch": "webpack --watch"
   },
   "keywords": [
     "security",
@@ -156,28 +159,35 @@
     "@typescript-eslint/eslint-plugin": "2.18.0",
     "@typescript-eslint/parser": "^2.0.0",
     "ajv": "^6.12.6",
+    "@webpack-cli/generators": "^2.0.0",
     "cross-spawn": "^6.0.5",
     "danger": "^10.6.4",
     "depcheck": "^1.4.1",
     "eslint": "6.8.0",
     "eslint-config-prettier": "^6.1.0",
     "fs-extra": "^9.1.0",
+    "file-loader": "^6.2.0",
     "jest": "^26.6.3",
     "jest-junit": "^12.0.0",
     "lodash": "^4.17.20",
+    "json-loader": "^0.5.7",
     "madge": "^3.4.4",
     "mock-fs": "^4.13.0",
     "nock": "^10.0.6",
     "npm-run-all": "^4.1.5",
     "prettier": "^1.18.2",
     "proxyquire": "^1.7.4",
+    "raw-loader": "^4.0.2",
     "restify": "^8.5.1",
     "sinon": "^4.0.0",
     "tap": "^12.6.1",
     "tap-junit": "^4.2.0",
     "ts-jest": "^26.5.2",
+    "ts-loader": "^9.0.2",
     "ts-node": "^8.0.0",
-    "typescript": "^3.4.1"
+    "typescript": "^3.9.9",
+    "webpack": "^5.34.0",
+    "webpack-cli": "^4.6.0"
   },
   "repository": {
     "type": "git",
@@ -185,13 +195,12 @@
   },
   "pkg": {
     "scripts": [
-      "dist/**/*.js"
+      "dist/**/*.js",
+      "bin"
     ],
     "assets": [
       "config.default.json",
-      "test-unpublished.json",
-      "help/**/*.txt",
-      "dist/STANDALONE",
+      "dist",
       "node_modules/@deepcode/dcignore"
     ]
   },
