@@ -1,4 +1,4 @@
-import { v4 as uuidv4 } from 'uuid';
+const { v4: uuidv4 } = require('uuid');
 const debug = require('debug')('snyk');
 const diff = require('diff');
 const exec = require('child_process').exec;
