@@ -1,13 +1,14 @@
-import * as tap from 'tap';
-import { only, test } from 'tap';
-import { fakeServer } from './fake-server';
-import * as cli from '../../src/cli/commands';
-
 const port = (process.env.PORT = process.env.SNYK_PORT = '12345');
 const BASE_API = '/api/v1';
 process.env.SNYK_API = 'http://localhost:' + port + BASE_API;
 process.env.SNYK_HOST = 'http://localhost:' + port;
 process.env.LOG_LEVEL = '0';
+
+import * as tap from 'tap';
+import { only, test } from 'tap';
+import { fakeServer } from './fake-server';
+import * as cli from '../../src/cli/commands';
+
 const apiKey = '123456789';
 let oldkey;
 let oldendpoint;
