@@ -11,7 +11,7 @@ import {
   generateScanResults,
 } from './results-formatter.fixtures';
 import { issuesToLineNumbers } from '@snyk/cloud-config-parser';
-import { PolicyMetadata } from '../../../../dist/cli/commands/test/iac-local-execution/types';
+import { PolicyMetadata } from '../../../../src/cli/commands/test/iac-local-execution/types';
 
 jest.mock('@snyk/cloud-config-parser');
 
